requirements.txt
Flask==2.0.1
apscheduler==3.7.0
requests==2.25.1
python-dotenv==0.17.1